+++ templates/content.php
@@ -12,7 +12,7 @@
       <h2><a href="<?php the_permalink(); ?>"><?php the_title(); ?></a></h2>
       <?php get_template_part('templates/entry-meta'); ?>
     </header>
-    <div class="entry-content">
+    <div class="entry-summary">
       <?php the_excerpt(); ?>
     </div>
     <footer>