+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### HEAD
+* Update to jQuery 1.8.3
 * Use `entry-summary` class for excerpts per Readability's Article Publishing Guidelines
 * Cleanup/refactor `lib/activation.php`
 * Remove `lib/post-types.php` and `lib/metaboxes.php`